third_party/blink/web_tests/wpt_internal/prerender/restriction-presentation-request.https.html

<!DOCTYPE html>
<title>Same-origin prerendering cannot start presentation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-channel');

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  // Start prerendering a page that attempts to start presentation.
  startPrerendering(`resources/presentation-request.html`);

  const result = await gotMessage;
  assert_equals(result,
    "Failed to execute 'start' on 'PresentationRequest': PresentationRequest::start() requires user gesture."
    );
}, 'prerendering page cannot start presentation');

</script>
</body>
